requirements.txt
# Core MIFARE functionality
pyscard>=2.0.7
pycryptodome>=3.19.0
colorama>=0.4.6
click>=8.1.7

# Web framework and API
flask>=2.3.3
flask-sqlalchemy>=3.0.5
flask-login>=0.6.3
flask-wtf>=1.1.1
flask-mail>=0.9.1
flask-cors>=4.0.0

# Database
sqlite3

# Security and authentication
werkzeug>=2.3.7
bcrypt>=4.0.1
PyJWT>=2.8.0

# QR code generation
qrcode>=7.4.2
Pillow>=10.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
email-validator>=2.0.0
